Conflicts:
	structr-core/src/main/java/org/structr/core/entity/AbstractNode.java
	structr-ui/src/main/resources/structr/js/schema.js

+++ structr-core/src/main/java/org/structr/core/entity/LinkedTreeNode.java
@@ -206,7 +206,7 @@ public abstract class LinkedTreeNode<R extends AbstractChildren<T, T>, S extends
 
 			Integer pos = rel.getProperty(positionProperty);
 
-			if (pos != null && pos.intValue() == position) {
+			if (pos != null && pos == position) {
 
 				return (T) rel.getTargetNode();
 			}
@@ -223,7 +223,7 @@ public abstract class LinkedTreeNode<R extends AbstractChildren<T, T>, S extends
 			Integer pos = rel.getProperty(positionProperty);
 			if (pos != null) {
 
-				return pos.intValue();
+				return pos;
 			}
 		}
